Add files via upload
Brief summary of My interests and intent

according to investopedia,the term unicorn refers to a privately held startup company with a value of over $1 billion. It is commonly used in the venture capital industry. The term was first popularized by venture capitalist Aileen Lee. Unicorns are very rare and require innovation. Because of their sheer size, unicorn investors tend to be private investors or venture capitalists, which means they are out of the reach of retail investors. Although it isn't necessary, many unicorns work their way to going public.
 
 My project is to uncover formation patterns,locations and distributions accross the world in unicorn industries. I will examine relationships between industry type and location,the number of USA unicorns and their distribution by cities,based on unicorn valutions,i will establish the correlation between valuation, and number of unicorn in a location,i will calculate the city or region's valuations and establish trends in the unicorn industry and many related questions, as the my datamay permit .